adjusting next order price preview
