+++ sections/futures/MarketInfo/MarketInfo.tsx
@@ -20,7 +20,6 @@ type MarketInfoProps = {
 	position: FuturesPosition | null;
 	openOrders: any[];
 	potentialTrade: PotentialTrade | null;
-	orderType: number;
 	refetch(): void;
 };
 
@@ -30,7 +29,6 @@ const MarketInfo: FC<MarketInfoProps> = ({
 	openOrders,
 	refetch,
 	potentialTrade,
-	orderType,
 }) => {
 	const { t } = useTranslation();
 	const { useExchangeRatesQuery } = useSynthetixQueries();
@@ -67,7 +65,6 @@ const MarketInfo: FC<MarketInfoProps> = ({
 				openOrders={openOrders}
 				refetch={refetch}
 				potentialTrade={potentialTrade}
-				orderType={orderType}
 			/>
 		</Container>
 	);
