+++ sections/futures/MarketInfoBox/MarketInfoBox.tsx
@@ -1,13 +1,18 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import styled from 'styled-components';
 import Wei, { wei } from '@synthetixio/wei';
 import InfoBox from 'components/InfoBox';
 import { formatCurrency, formatPercent, zeroBN } from 'utils/formatters/number';
 import { CurrencyKey, Synths } from '@synthetixio/contracts-interface';
-import { PotentialTrade } from '../types';
+import { PositionSide, PotentialTrade } from '../types';
 import PreviewArrow from 'components/PreviewArrow';
 import useGetFuturesPotentialTradeDetails from 'queries/futures/useGetFuturesPotentialTradeDetails';
 import { FuturesPosition, FuturesPotentialTradeDetails } from 'queries/futures/types';
+import useGetNextPriceDetails from 'queries/futures/useGetNextPriceDetails';
+import { computeNPFee } from 'utils/costCalculations';
+import { getMarketKey } from 'utils/futures';
+import { useRecoilValue } from 'recoil';
+import { networkState } from 'store/wallet';
 
 type MarketInfoBoxProps = {
 	position: FuturesPosition | null;
@@ -15,6 +20,9 @@ type MarketInfoBoxProps = {
 	potentialTrade: PotentialTrade | null;
 	marketMaxLeverage: Wei | undefined;
 	currencyKey: string;
+	tradeSize: Wei;
+	side: PositionSide;
+	orderType: number;
 };
 
 const MarketInfoBox: React.FC<MarketInfoBoxProps> = ({
@@ -23,7 +31,11 @@ const MarketInfoBox: React.FC<MarketInfoBoxProps> = ({
 	potentialTrade,
 	marketMaxLeverage,
 	currencyKey,
+	tradeSize,
+	side,
+	orderType,
 }) => {
+	const network = useRecoilValue(networkState);
 	const totalMargin = position?.remainingMargin ?? zeroBN;
 	const availableMargin = position?.accessibleMargin ?? zeroBN;
 	const buyingPower =
@@ -42,6 +54,25 @@ const MarketInfoBox: React.FC<MarketInfoBoxProps> = ({
 	);
 
 	const previewTrade = potentialTradeDetails.data ?? null;
+	const isNextPriceOrder = orderType === 1;
+	const nextPriceDetailsQuery = useGetNextPriceDetails(getMarketKey(currencyKey, network.id));
+	const nextPriceDetails = nextPriceDetailsQuery?.data;
+
+	const positionSize = position?.position?.size ? wei(position?.position?.size) : zeroBN;
+	const orderDetails = useMemo(() => {
+		const newSize = side === PositionSide.LONG ? tradeSize : -tradeSize;
+		return { newSize, size: (positionSize ?? zeroBN).add(newSize).abs() };
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [side, positionSize]);
+
+	const { commitDeposit } = useMemo(
+		() => computeNPFee(nextPriceDetails, wei(orderDetails.newSize)),
+		[nextPriceDetails, orderDetails]
+	);
+
+	const totalDeposit = useMemo(() => {
+		return (commitDeposit ?? zeroBN).add(nextPriceDetails?.keeperDeposit ?? zeroBN);
+	}, [commitDeposit, nextPriceDetails?.keeperDeposit]);
 
 	const getPotentialAvailableMargin = (
 		previewTrade: FuturesPotentialTradeDetails | null,
@@ -61,21 +92,28 @@ const MarketInfoBox: React.FC<MarketInfoBoxProps> = ({
 		return previewTrade?.margin?.sub(inaccessible).abs() ?? zeroBN;
 	};
 
+	const previewAvailableMargin = React.useMemo(() => {
+		const potentialAvailableMargin = getPotentialAvailableMargin(previewTrade, marketMaxLeverage);
+		return isNextPriceOrder
+			? potentialAvailableMargin?.sub(totalDeposit) ?? zeroBN
+			: potentialAvailableMargin;
+	}, [previewTrade, marketMaxLeverage, isNextPriceOrder, totalDeposit]);
+
 	const previewTradeData = React.useMemo(() => {
 		const size = wei(potentialTrade?.size || zeroBN);
-		const potentialAvailableMargin = getPotentialAvailableMargin(previewTrade, marketMaxLeverage);
-		const potentialBuyingPower = potentialAvailableMargin?.mul(marketMaxLeverage).abs() ?? zeroBN;
-		const enumeratorMarginUsage = previewTrade?.margin?.sub(potentialAvailableMargin);
-		const potentialMarginUsage = enumeratorMarginUsage?.div(previewTrade?.margin)?.abs() ?? zeroBN;
+		const potentialMarginUsage =
+			previewTrade?.margin?.sub(previewAvailableMargin)?.div(previewTrade?.margin)?.abs() ?? zeroBN;
+		const potentialBuyingPower =
+			previewAvailableMargin?.mul(marketMaxLeverage ?? zeroBN)?.abs() ?? zeroBN;
 
 		return {
 			showPreview: size && !size.eq(0),
 			totalMargin: previewTrade?.margin || zeroBN,
-			availableMargin: potentialAvailableMargin.gt(0) ? potentialAvailableMargin : zeroBN,
+			availableMargin: previewAvailableMargin.gt(0) ? previewAvailableMargin : zeroBN,
 			buyingPower: potentialBuyingPower.gt(0) ? potentialBuyingPower : zeroBN,
-			marginUsage: potentialMarginUsage.gt(0) ? potentialMarginUsage : zeroBN,
+			marginUsage: potentialMarginUsage.gt(1) ? wei(1) : potentialMarginUsage,
 		};
-	}, [potentialTrade?.size, previewTrade, marketMaxLeverage]);
+	}, [potentialTrade?.size, previewTrade?.margin, previewAvailableMargin, marketMaxLeverage]);
 
 	return (
 		<StyledInfoBox
