+++ sections/futures/MarketInfoBox/index.ts
@@ -1,2 +1,1 @@
-export { default as MarketInfoBox } from './MarketInfoBox';
-export { default as NextPriceInfoBox } from './NextPriceInfoBox';
+export { default } from './MarketInfoBox';
